refactor: Moving QCCheck & QCChecVar around
